artiko/requirements.txt
Flask>=2.2.0
gunicorn
python-dotenv
